pyproject.toml
[project]
name = "extract_rag"
version = "0.0.1"
description = "Starter template for making a custom RAG research agent graph in LangGraph."
authors = [
    { name = "Vadym Barda", email = "19161700+vbarda@users.noreply.github.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.6",
    "langchain>=0.2.14",
    "langchain-fireworks>=0.1.7",
    "python-dotenv>=1.0.1",
    "msgspec>=0.18.6",
    "langchain-mongodb>=0.1.9",
    "langchain-google-genai>=1.0.10",
    "langchain-community>=0.2.19",
    "langchain-qdrant>=1.0.0",
    "langgraph-cli>=0.4.7",
    "langgraph-api>=0.5.25",
    "langchain-experimental>=0.4.0",
    "fastembed>=0.7.3",
    "qdrant-client==1.15.1",
    "langextract>=1.1.0",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["retrieval_graph", "shared"]
[tool.setuptools.package-dir]
"langgraph.templates.retrieval_graph" = "src/retrieval_graph"
"retrieval_graph" = "src/retrieval_graph"
"shared" = "src/shared"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"
[tool.pytest.ini_options]
pythonpath = [
  "src"
]
